# remotion-composer/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMontage — Talking-head auto-edit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "summary": "List Jobs",
        "description": "Every build, from the legacy root and from inside every project.\n\nOne list rather than two so a legacy job stays openable at the same URL it\nalways had — that is the point of reading both layouts instead of migrating.",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummaryResponse"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/events": {
      "get": {
        "summary": "Stream Events",
        "operationId": "stream_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/chat": {
      "get": {
        "summary": "Chat History",
        "operationId": "chat_history_api_jobs__job_id__chat_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat History Api Jobs  Job Id  Chat Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Chat Turn",
        "description": "One conversational revise turn. See `_run_chat_turn` for the details.",
        "operationId": "chat_turn_api_jobs__job_id__chat_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs": {
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/revise": {
      "post": {
        "summary": "Revise Run",
        "description": "Reuses the exact chat-turn code `POST /jobs/{id}/chat` runs — see\n`api_jobs._run_chat_turn`. The one thing an automation-created run adds:\nif it was created with `render=true`, a revise re-queues render too.\n\nDefaults to False, not True: a job that never went through `POST\n/api/runs` (created via the UI or `POST /api/jobs`) has no `render` option\nat all, and re-rendering by default on a job nobody asked this endpoint to\nmanage would be a surprise render most callers of a UI-created job did not\nask for.",
        "operationId": "revise_run_api_runs__run_id__revise_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviseRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "cancelled"
        ],
        "title": "CancelRunResponse"
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "turn": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied",
            "default": false
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "preview_queued": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Queued"
          },
          "remedy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remedy"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "options_changed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options Changed"
          },
          "not_done": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Not Done"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ChatResponse"
      },
...
      "CreateJobRequest": {
        "properties": {
          "input_path": {
            "type": "string",
            "title": "Input Path",
            "default": ""
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "stages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stages"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "CreateJobRequest"
      },
      "CreateJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "queue_position": {
            "type": "integer",
            "title": "Queue Position"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "job_id",
          "queue_position"
        ],
        "title": "CreateJobResponse"
      },
...
      "CreateRunRequest": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/RunSource"
            },
            "type": "array",
            "minItems": 1,
            "title": "Sources"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "render": {
            "type": "boolean",
            "title": "Render",
            "default": true
          },
          "render_scale": {
            "type": "number",
            "title": "Render Scale",
            "default": 1.0
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "sources"
        ],
        "title": "CreateRunRequest"
      },
      "CreateRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "status_url",
          "events_url"
        ],
        "title": "CreateRunResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobSummaryResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending"
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "default": 0.0
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version",
            "default": 0
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "stages": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stages"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionEntry"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobSummaryResponse"
      },
...
      "ReviseRunRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ReviseRunRequest"
      },
      "ReviseRunResponse": {
        "properties": {
          "turn": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied",
            "default": false
          },
          "status_url": {
            "type": "string",
            "title": "Status Url",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ReviseRunResponse"
      },
...
      "RunError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "RunError"
      },
...
      "RunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "awaiting_render"
            ],
            "title": "State"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "percent": {
            "type": "integer",
            "title": "Percent",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunError"
              },
              {
                "type": "null"
              }
            ]
          },
          "outputs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOutputs"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version",
            "default": 0
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionEntry"
            },
            "type": "array",
            "title": "Versions"
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "default": 0.0
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At",
            "default": 0.0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "state"
        ],
        "title": "RunStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}